Rework tone map to use custom function.
The im_maplut function in vips is not exactly what I want, so implement
a closer match by hand. In the process, set down the definition of the
output result parameters that I really want.

Now that we have a better tone map function, remove the display gamma
controls. They are superfluous.